Add useIsLargeScreen hook to replace ScreenSizeProvider
This is a more native way to determine if the screen size is large, and is less prone to errors than the old manual approach was.
