+++ src/root/hooks/index.ts
@@ -1,2 +1,3 @@
 export { default as useDebounce } from "src/root/hooks/useDebounce";
 export { default as useHash } from "src/root/hooks/useHash";
+export { default as useIsLargeScreen } from "src/root/hooks/useIsLargeScreen";
