+++ src/root/index.ts
@@ -1,5 +1,4 @@
 export * from "src/root/components";
 export * from "src/root/hooks";
-export * from "src/root/providers";
 
 export type * from "src/root/types";
