--- src/root/providers/index.ts
@@ -1,10 +1,0 @@
-export {
-  default as ScreenSizeProvider,
-  useScreenSize,
-} from "src/root/providers/ScreenSizeProvider";
-
-export type {
-  ScreenSizeProps,
-  ScreenSizeContextValue,
-  ScreenDimensions,
-} from "src/root/providers/ScreenSizeProvider";
